requirements.txt
Django>=4.0,<4.1
wagtail>=3.0,<3.1
gunicorn==20.1.0
wagtailsvg
django-inline-svg==0.1.1
django-environ==0.9.0
psycopg2-binary==2.9.3


django-extensions==3.1.5
ipython==8.4.0
django-tailwind==3.1.1


